requirements.txt
requests==2.32.3
selenium==4.26.1
pytest==8.3.3
allure-pytest==2.13.5
Pillow==11.0.0




